+++ src/Properties/resources/database/migrations/2023_08_23_155708_add_hidden_field_to_properties.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class () extends Migration {
+    public function up(): void
+    {
+        Schema::table('properties', function (Blueprint $table) {
+            $table->boolean('is_hidden')->default(false)->after('configuration');
+        });
+    }
+
+    public function down(): void
+    {
+        Schema::table('properties', function (Blueprint $table) {
+            $table->dropColumn('is_hidden');
+        });
+    }
+};